requirements.txt
streamlit==1.13.0
opencv-python==4.6.0.66
tensorflow==2.10.0
Pillow==9.2.0
numpy==1.23.1
image==1.5.33